break out component tests to their own files
